% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlq_lmc21.R
\name{qlq_lmc21}
\alias{qlq_lmc21}
\title{Score the EORTC QLQ-LMC21 Quality of Life Questionnaire}
\usage{
qlq_lmc21(df, items = NULL, keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame containing responses to the 21 QLQ-LMC21 items, and
possibly other variables.}

\item{items}{A character vector with the QLQ-LMC21 item names, or a numeric
vector indicating the column numbers of the QLQ-LMC21 items in \code{df}.
If \code{items} is omitted, then \code{qlq_lmc21} will assume that
\code{df} contains \strong{ONLY} the QLQ-LMC21 items and no other variables.
See Details for more information.}

\item{keepNvalid}{Logical, whether to return variables containing the
number of valid, non-missing items on each scale for each respondent should
be returned in the data frame with the scale scores.  The default is
\code{FALSE}.  Set to \code{TRUE} to return these variables, which will be
named \code{"scalename_N"} (e.g., \code{QL_N}).  Most users should omit
this argument entirely.  This argument might be removed from future
versions of the package, so please let me know if you think this argument
useful and would rather it remain a part of the function.}
}
\value{
A data frame with all 13 of the QLQ-LMC21 scores is returned.  Of the 13
scores, all 13 are Symptom Scales (see below).  Of
the 13 Symptom Scales, 9 are based on a single item each.  All scores are
scaled to range from 0-100, even scores based on single items.  Be aware that
these single-item scales still have only 4 possible values, even though they
are transformed to range from 0-100.  The scale names and numbers of items are
listed below.

\strong{Symptom Scales (higher is more symptoms, worse functioning)}
\itemize{
\item \strong{LMC_NP} - Nutritional problems (from 2 items)
\item \strong{LMC_FA} - Fatigue (from 3 items)
\item \strong{LMC_PA} - Pain (from 3 items)
\item \strong{LMC_EP} - Emotional problems (from 4 items)
\item \strong{LMC_WL} - Weight loss (from 1 item)
\item \strong{LMC_TA} - Taste (from 1 item)
\item \strong{LMC_DM} - Dry mouth (from 1 item)
\item \strong{LMC_SM} - Sore mouth/tongue (from 1 item)
\item \strong{LMC_PN} - Peripheral neuropathy (from 1 item)
\item \strong{LMC_JA} - Jaundice (from 1 item)
\item \strong{LMC_FR} - Contact with friends (from 1 item)
\item \strong{LMC_FE} - Talking about feelings (from 1 item)
\item \strong{LMC_SX} - Sex life (from 1 item)
}

Optionally, the data frame can additionally have variables containing the
number of valid item responses on each scale for each respondent (if
\code{keepNvalid = TRUE}, but this option might be removed in future package
updates).
}
\description{
Scores the European Organization for Research and Treatment of
Cancer (EORTC) QLQ-LMC21 Colorectal Liver Cancer Module.
}
\details{
This function returns a total of 13 different scores from the EORTC
QLQ-LMC21. Scores are calculated according to the official scoring algorithms
from the EORTC.

In addition to the name of your data frame containing the QLQ-LMC21 item
responses (\code{df}), you need to tell the function how to find the
variables that correspond to the QLQ-LMC21 items in \code{df}.  You can do this
in 1 of 2 ways:
\enumerate{
\item The first way is to manually provide the item names or locations
using the \code{items} argument.  For example, if your first 10
variables in \code{df} contain demographics, followed by the 21 QLQ-LMC21
items \strong{in order} starting with the 11th variable, then you could
use \code{items = 11:36}.
\item The second way only applies if your data frame (\code{df}) contains
\strong{ONLY} the 21 variables corresponding to the 21 QLQ-LMC21 items,
in order, with no other non-QLQ-LMC21 variables.  In this case, you can
just use the \code{df} argument and omit \code{items}.
}
}
\section{How Missing Data is Handled}{

The \code{qlq_lmc21} function will calculate the scale scores as long as at
least half of the items on the given scale have valid, non-missing item
responses.  Scores calculated in the presence of missing
items are pro-rated so that their theoretical minimum and maximum values
are identical to those from scores calculated from complete data.
}

\examples{

\dontrun{
dat <- PROscorerTools::makeFakeData(n = 10, nitems = 21, prefix = "lmc", values = 1:4)
qlq_lmc21(dat, items = 1:21)
}
}
\references{
Blazeby JM, Fayers P, Conroy T, et al. Validation of the EORTC QLQ-LCM21
Questionnaire for Assessment of Patient-Reported Outcomes During Treatment of
Colorectal Liver Metastases. \emph{Br J Surg 96}:291-298, 2009.

Kavadas V, Blazeby JM, et al.  Development of an EORTC disease-specific
quality of life questionnaire for use in patients with liver metastases from
colorectal cancer. \emph{Eur J Cancer}. 2003 Jun;39(9):1259-63.
}
