% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFSFI.R
\name{scoreFSFI}
\alias{scoreFSFI}
\title{Score the Female Sexual Function Index (FSFI)}
\usage{
scoreFSFI(df, iprefix = "fsfi", keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame containing responses to the 19 FSFI items, and
possibly other variables.}

\item{iprefix}{Item number prefix.  Quote the letter(s) preceding the FSFI
item numbers in your data frame.  If this argument is omitted, the function
will assume that your items are named "fsfi1", "fsfi2", etc.}

\item{keepNvalid}{Logical, whether to return variables containing the
number of valid, non-missing items on each scale for each respondent should
be returned in the data frame with the scale scores.  The default is
\code{FALSE}.  Set to \code{TRUE} to return these variables, which will be
named \code{"scalename_N"} (e.g., \code{fsfi_pain_N}).}
}
\value{
A data frame with the following scale scores is returned:

\itemize{
  \item \strong{fsfi_des} - FSFI Desire subscale (range 2 - 6)
  \item \strong{fsfi_arous} - FSFI Arousal subscale
  \item \strong{fsfi_lub} - FSFI Lubrication subscale
  \item \strong{fsfi_org} - FSFI Orgasm subscale
  \item \strong{fsfi_sat} - FSFI Satisfaction subscale
  \item \strong{fsfi_pain} - FSFI Pain subscale
  \item \strong{fsfi_tot} - FSFI Total score
  \item \strong{fsfi_dys01} - Indicator of FSFI sexual dysfunction (i.e.,
    fsfi_tot <= 26.55)
  \item \strong{fsfi_nzero15} - There are 15 FSFI items that have a response
    option of \code{0} ("No sexual activity").  This is the number of those
    items with responses of 0 or NA. (See Details.)
  \item \strong{fsfi_sexactive01} - For the FSFI scores to be valid estimates
    of sexual functioning, respondents need to have been sexually active
    during the 4 week recall period.  This variable indicates whether their
    sexual activity levels were high enough for their FSFI scores to be
    valid. Specifically, it is an indicator that fsfi_nzero15 <= 7.  (See
    Details.)
}

  Optionally, the data frame can additionally have variables containing the
  number of valid item responses on each scale for each respondent.
}
\description{
Scores the Female Sexual Function Index (FSFI)
}
\examples{
dat <- PROscorerTools::makeFakeData(n = 10, nitems = 19, values = 0:5)
missFix <- function(df, missvals) {
  fixer <- function(x) {
    x[x \%in\% missvals] <- NA
    x
  }
  df[] <- lapply(df, fixer )
  df
}
dat[c(1, 2, 15, 16)] <- missFix(dat[c(1, 2, 15, 16)], 0)
names(dat) <- paste0("f", 1:19)
scoreFSFI(dat, "f")
}

