% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreNARCC.R
\name{scoreNARCC}
\alias{scoreNARCC}
\title{Score the Cognitive Causation (CC) and Negative Affect in Risk (NAR)
  scales}
\usage{
scoreNARCC(df, items = NULL, whichScale, minmax = c(0, 3), okmiss = 0.5,
  keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame containing responses to the CC and/or NAR items, and
possibly other variables.}

\item{items}{(optional) A character vector with the CC or NAR item names, or
a numeric vector indicating the column numbers of the CC or NAR items in
\code{df}.  If \code{items} is omitted, then \code{scoreNARCC} will assume
that \code{df} contains only the items to be scored (either CC or NAR
items) and no non-scored variables.}

\item{whichScale}{Either \code{"CC"} or \code{"NAR"}, the scale you wish to
score.}

\item{minmax}{A vector of 2 integers with format \code{c(itemMin, itemMax)},
indicating the minimum and maximum possible item responses.  The default
value is \code{c(0, 3)}, and assumes that the item responses are coded from
0 to 3.  If, instead, your item responses are coded from 1 to 4, then enter
\code{c(1, 4)} for this argument.}

\item{okmiss}{The maximum proportion of items on \code{whichScale} that a
respondent is allowed to have missing and still have their non-missing
items scored (and prorated). If the proportion of missing items for a
respondent is greater than \code{okmiss}, then the respondent will be
assigned a value of \code{NA} for their scale score.  The default value is
\code{0.50}, and this generally should not be changed.}

\item{keepNvalid}{Logical value indicating whether a variable containing the
number of valid, non-missing items for each respondent should be returned
in a data frame with the scale score.  The default is \code{FALSE}.  Set to
\code{TRUE} to return this variable, which will be named
\code{"whichScale_N"} (with whatever name you gave to the \code{whichScale}
argument).}
}
\description{
Scores the Cognitive Causation (CC) and Negative Affect in Risk
  (NAR) scales, two scales measuring intuitive elements of cancer risk
  perception.
}
\details{

}
\examples{
# Make fake data for the example
nardat <- PROscorerTools::makeFakeData(nitems = 6, values = 0:3,
                                       propmiss = 0.40, prefix = "nar")
ccdat <- PROscorerTools::makeFakeData(nitems = 7, values = 0:3,
                                      propmiss = 0.40, prefix = "cc",
                                      id = TRUE)

# The nardat data frame contains ONLY NAR items, so can omit "items" argument
scoreNARCC(nardat, whichScale = "NAR")

# The ccdat data frame contains an "ID" variable, so need to use "items" arg
names(ccdat)

# The "items" argument can be either the numeric vector indexing
#     the location of the items in df,
#     or a character vector of the item names
scoreNARCC(ccdat, items = 2:8, whichScale = "CC")

cc_names <- c("cc1", "cc2", "cc3", "cc4", "cc5", "cc6", "cc7")
scoreNARCC(ccdat, items = cc_names, whichScale = "CC")
}
\references{
Cohen, P, Cohen, J, Aiken, LS, & West, SG. (1999). The problem of units and
the circumstance for POMP. \emph{Multivariate Behavioral Research}, 34(3),
315–346.
}

